Adding Support for 'BigDecimal' for integer, long, double, string type. (#4395)
Adding Support for 'BigDecimal' for the following 'Target type' - integer, long, double, string

+++ data-prepper-api/src/main/java/org/opensearch/dataprepper/typeconverter/DoubleConverter.java
@@ -14,7 +14,7 @@ public class DoubleConverter implements TypeConverter<Double> {
             return (Double)source;
         }
         if (source instanceof Number) {
-            return (double)(((Number)source).intValue());
+            return (((Number)source).doubleValue());
         }
         if (source instanceof Boolean) {
             return (double)(((Boolean)source) ? 1.0 : 0.0);
